Step1 of the PCARS DS data provider refactor.
- Added a service for all the stuuf that is specific to our app.
- Started refactoring the hook to be a generic object that handles
the dialogue wiht the PCARS DS server. Could be used by other apps in the future.
- Start of a generic API to communicate between both.
- Move some of the DB management and message broadcast from the hook to the service

Still much to move. For now I addressed all the part of the code that does not need
a running session on the server (connection and tracks/classes/vehicles DB update).
All the code that handle server responses with logs and session status still need to be divided/moved

State is semi-fonctional for now. Should be similar to previous state but with less error handling and messages.